+++ CHANGELOG.md
@@ -17,6 +17,7 @@ hideFromSearchEngines: true
 - Make `filename` optional for `CodeSnippetTitle`.
 - Use passive event listeners on `Search`, `NumberedCodeSnippet`, and `OnThisPage` to improve performance. [#437](https://github.com/mapbox/dr-ui/pull/437)
 - Make `Video` disable autoplay by default. [#276](https://github.com/mapbox/dr-ui/pull/276)
+- Load the `Search` component with a facade to improve performance. [#430](https://github.com/mapbox/dr-ui/pull/430)
 
 ## 3.3.1